% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers
A function to add to numbers with an unrelated example.}
\usage{
add(a, b)
}
\arguments{
\item{a}{Number}

\item{b}{Number}
}
\description{
Add two numbers
A function to add to numbers with an unrelated example.
}
\examples{
par(mar = c(5, 5, 5, 5), cex.lab = 1.3)
plot(1:10, xlab = expression(hat(beta)), ylab = expression(hat(beta)))
}
